Fixed command and parsing for inventorying running systemd services
The awk command just looks wrong to me.
^UNIT should match lines starting with UNIT.
The output of systemctl list-units on debian-11 has UNIT with some spaces at the beginning.
Also, later in the output some other footer elements are present.

Not sure when and where this systemctl | awk command worked but with debian-11, systemctl 247 it doesn't and --output=json is a nice option.